# packages/contracts/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/promotions/admin": {
      "get": {
        "operationId": "PromotionsAdminController_listPromotions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Promotions Admin"]
      }
    },
    "/api/v1/promotions/admin/{promotionId}": {
      "get": {
        "operationId": "PromotionsAdminController_getPromotion",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Promotions Admin"]
      }
    },
    "/api/v1/promotions/admin/{promotionKey}": {
      "put": {
        "operationId": "PromotionsAdminController_upsertPromotion",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Promotions Admin"]
      }
    },
    "/api/v1/promotions/admin/{promotionId}/coupons/{code}": {
      "put": {
        "operationId": "PromotionsAdminController_upsertCoupon",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Promotions Admin"]
      }
    },
...
    "/api/v1/orders/admin/{orderId}": {
      "get": {
        "operationId": "AdminOrdersController_getOrder",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Orders Admin"]
      }
    },
...
    "/api/v1/checkout/reserve": {
      "post": {
        "operationId": "CheckoutController_reserveCheckoutPreview",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Checkout"]
      }
    },
...
    "/api/v1/returns/admin/{returnRequestId}": {
      "get": {
        "operationId": "AdminReturnsController_getReturn",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Returns Admin"]
      }
    },
...
    "/api/v1/reviews/admin": {
      "get": {
        "operationId": "AdminReviewsController_listReviews",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Reviews Admin"]
      }
    },
    "/api/v1/reviews/admin/{reviewId}": {
      "get": {
        "operationId": "AdminReviewsController_getReview",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Reviews Admin"]
      }
    },
    "/api/v1/reviews/admin/{reviewId}/moderate": {
      "patch": {
        "operationId": "AdminReviewsController_moderateReview",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Reviews Admin"]
      }
    },
...
  },
  "info": {
    "title": "Universal E-Commerce API",
    "description": "Phase 00 foundation contract for the headless commerce backend.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}